Fix Android push registration upsert handling
